Move GA tags to bottom of head tag as it is loaded asynchronously
http://stackoverflow.com/questions/3173571/should-i-put-the-google-analytics-js-in-the-head-or-at-the-end-of-body
One of the main advantages of the asynchronous snippet is that you can position it at the top of the HTML document. This increases the likelihood that the tracking beacon will be sent before the user leaves the page. It is customary to place JavaScript code in the <head> section, and we recommend placing the snippet at the bottom of the section for best performance.