Add canonical tag for pages other than the home page
http://seositecheckup.com/seo-audit/www.corporateprogrammer.com#
The Canonical Link Tag can be used when there are several pages with similar content and you want to tell the search engines which page you prefer to use in the search results. If your webpage does not have duplicate content and it has the preferred URL you must remove the canonical link tag.