Fixed erroneous commits by adding tagged and all post display SCSS and added missing tagged-post include
I could have done a rebase and forced the push. I chose not to. Possibly because I was not in the right frame of mind. So now I have a lame commit. This may bite me in the future.
Find opportunities in the mistakes you make http://alistapart.com/article/finding-opportunities-in-the-mistakes-we-make
#YOLO